fix resetMeeting content
