+++ src/server/graphql/models/Team/endMeeting/resetMeeting.js
@@ -1,5 +1,8 @@
+import {convertFromRaw, convertToRaw} from 'draft-js';
 import getRethink from 'server/database/rethinkDriver';
 import {DONE, LOBBY} from 'universal/utils/constants';
+import addTagToProject from 'src/universal/utils/draftjs/addTagToProject';
+import getTagsFromEntityMap from 'src/universal/utils/draftjs/getTagsFromEntityMap';
 
 export default async function resetMeeting(teamId) {
   return new Promise((resolve) => {
@@ -10,17 +13,18 @@ export default async function resetMeeting(teamId) {
         .filter((project) => project('tags').contains('archived').not())
         .pluck('id', 'content', 'tags')
       const archivedProjects = projects.map((project) => {
-        const nextContent =
+        const contentState = convertFromRaw(JSON.parse(project.content));
+        const nextContentState = addTagToProject(contentState, '#archived');
+        const raw = convertToRaw(nextContentState);
+        const nextTags = getTagsFromEntityMap(raw.entityMap);
+        const nextContentStr = JSON.stringify(raw);
         return {
-          content: nextContent,
+          content: nextContentStr,
           tags: nextTags,
           id: project.id
         }
-      })
-        //.update((project) => ({
-        //  tags: project('tags').append('archived'),
-        //  content: project('content').add(' archived')
-        //}));
+      });
+      //const archivedProjectIds = projects.map((proj) => proj.id);
 
       await r.table('Team').get(teamId)
         .update({
@@ -30,6 +34,14 @@ export default async function resetMeeting(teamId) {
           facilitatorPhaseItem: null,
           meetingPhaseItem: null,
           activeFacilitator: null
+        })
+        .do(() => {
+          return r.args(archivedProjects).forEach((project) => {
+            return r.table('Project').update(project('id'), {
+              content: project('content'),
+              tags: project('tags')
+            })
+          })
         })
         .do(() => {
           // flag agenda items as inactive (more or less deleted)
@@ -48,12 +60,12 @@ export default async function resetMeeting(teamId) {
             .sample(100000)
             .coerceTo('array')
             .do((arr) => arr.forEach((doc) => {
-              return r.table('TeamMember').get(doc('id'))
+                return r.table('TeamMember').get(doc('id'))
                   .update({
                     checkInOrder: arr.offsetsOf(doc).nth(0),
                     isCheckedIn: null
                   });
-            })
+              })
             );
         });
       resolve();
